requirements.txt
python-dotenv
requests
venv\Scripts\activate
